Implemented the following changes: Added error checking to see if the hcp is of type OB/GYN when clicking the edit or add buttons.
Added checking to make sure it was impossible to add more than one
current pregnancy.

Added a drop-down for selecting a date of a prior pregnancy information
to edit. Added methods to redirect to the editPriorPregnancy page and
populated the boxes with the correct data from the database. I have not
yet enabled the save button on editing prior pregnancies. It throws a
500 error now.