See below for changes.
Peer Programming with hmukhta, ffamator, jwmorris

We fixed both the front end and the backend. We are now able to edit a
prior pregnancy, current pregnancy, and add a current pregnancy.

Added an error message that shows when there are no prior pregnancies
available to edit.

We still need to implement input validation, unit tests, and cucumber
tests.