Skeleton Structure Creation
I have continued to add on to the front end skeleton. I created input
boxes for the data field. This is not a definitive decision on where
everything will be going, it is more to show how the space can be used
and to know what all needs to be shown on the pages.

I also created a controller class for the Obstetrics visit and a class
for the obstetrics form information. These are both empty.