Changes detailed below per file
I have added many of the methods that we will need to complete this
iteration. Almost all of them are just skeletons of the method and do
not actually perform a functionality. I have commented out places in
which the backend is being referred to so that there are no compile
errors.

I started adding some of the methods to the front end. I also created
more conditional rendering in the main landing page for obstetrics
office visits. These check for OBGYN and for a valid obstetrics patient.