Changes detailed below:
In the controller I fixed the method to check if a patient is a valid
obstetrics patient

In the form class, I created the variables and getter/setters we need.
Based it off the officeVistiForm class iTrust has. I also made the
submit buttons for the visit infor and fetal infor