Changes detailed below
I have added a validator that could be used to test the file type that
the user attempts to upload. We may or may not end up using this.

I added a method to the Navigation Controller that will redirect to the
obstetrics visit page.

I started working on the controller and the obstetrics visit form to add
some functionality to it. 

I also fixed some issues in the front end.