Ultrasound table
I added a table that will allow the HCP to see the names of the files
that they have already uploaded to that office visit. 

I edited the method in the MySQL class to return a list instead of a
single value since the SQL query was going to return 1+ any way. 

I also fixed the MySQL file tests to make sure that they are working
with the new List return type