Adding Read Only View & Logging
I have added a read only view of an obstetrics office visit information
that is seen by a regular HCP when they click the view button next to a
prior visit. I also added a button that will take an hcp to chart a
regular office visit instead of an obstetrics one since they are not
able to.

I also implemented logging for viewing, creating, editing obstetrics
office visits, and adding ultrasounds.