Added a link between Initialization and office visit
I added a linking field between the obstetrics initialization and the
obstetrics office visits so that we can tie the two together. This will
be very useful on UC 96.

I also made it so that the images have the time stamp appended to their
name so that we can ensure that they are unique.