Added new patient functionality
I added the functionality so that when a new baby is created there is a
new patient that is added to the itrust system.

I also fixed an issue with including primefaces where jquery was
conflicting. I solved it by making sure to include the primefaces jquery
in the file if there was no p tag being used.