Complications
I have added the ability to display complications based on the office
visits. Some of the complications are not based on an office visit and
are just indicators so I do not calculate them per office visit. If any
office visit for that pregnancy has a complication associated with them
then the complication will be set for the whole pregnancy.