Added Report Functionality
Linked the Diagnosis and Allergy information of a patient to what is
displayed in the report. The diagnosis/allergies that are reported are
those that were specified. Chronic Illnesses are now selected and
displayed in the report based on if their "is_chronic" flag is set.
Also, had a problem with duplicate diagnosis reports, those are now
fixed.